refactor: rename controller file
